refactor question types, and add a better help input
